requirements.txt
uvicorn==0.32.0
fastapi==0.115.4
requests==2.28.1
torch==2.4.0
torchvision==0.19
transformers==4.44.2
pydantic==2.9.2
pathlib==1.0.1
python-dotenv==1.0.1
sentencepiece==0.2.0
